+++ CHANGELOG.md
@@ -11,6 +11,8 @@
   See the [README][] for more info on [inheritable
   observables](https://github.com/moll/js-concert#inheriting).
 
+- Throws `TypeError` when calling `on` or `once` without a listener function.
+
 [README]: https://github.com/moll/js-concert
 
 ## 1.2.0 (Sep 29, 2014)